improve behavior of __dirname for non-browser targets
