+++ src/utils/morph.js
@@ -1,0 +1,12 @@
+module.exports = morph;
+
+// replace object properties
+function morph(object, newProperties) {
+  for (let key in object) {
+    delete object[key];
+  }
+
+  for (let key in newProperties) {
+    object[key] = newProperties[key];
+  }
+}
