+++ src/visitors/env.js
@@ -1,5 +1,6 @@
 const types = require('babel-types');
 const matchesPattern = require('./matches-pattern');
+const morph = require('../utils/morph');
 
 module.exports = {
   MemberExpression(node, asset) {
@@ -15,14 +16,3 @@ module.exports = {
     }
   }
 };
-
-// replace object properties
-function morph(object, newProperties) {
-  for (let key in object) {
-    delete object[key];
-  }
-
-  for (let key in newProperties) {
-    object[key] = newProperties[key];
-  }
-}
